+++ src/main/java/ru/romania/taco/web/DesignTacoController.java
@@ -1,4 +1,4 @@
-package ru.romania.taco;
+package ru.romania.taco.web;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -12,7 +12,12 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import ru.romania.taco.Ingredient;
+import ru.romania.taco.Taco;
+import ru.romania.taco.TacoOrder;
 import ru.romania.taco.Ingredient.Type;
+import org.springframework.web.bind.annotation.PostMapping;
+
 
 @Slf4j
 @Controller
@@ -38,7 +43,7 @@ public class DesignTacoController {
         Type[] types = Ingredient.Type.values();
         for (Type type : types) {
             model.addAttribute(type.toString().toLowerCase(), filterByType(ingredients, type));
-
+            filterByType(ingredients, type);
         }
     }
 
@@ -57,9 +62,17 @@ public class DesignTacoController {
         return "design";
     }
 
+    @PostMapping()
+    public String processTaco(Taco taco, @ModelAttribute TacoOrder tacoOrder) {
+        tacoOrder.addTaco(taco);
+        log.info("Processing taco: {}", taco);
+        return "redirect:/orders/current";
+    }
+    
+
     private Iterable<Ingredient> filterByType (List<Ingredient> ingredients, Type type) {
         return ingredients.stream()
-                .filter(x -> x.type().equals(type))
+                .filter(x -> x.getType().equals(type))
                 .collect(Collectors.toList());
     }
 }
